Solution to Calculate and Print the Area of a Circle

+++ Calculate and Print the Area of a Circle@@6788d5012db25efec946f943/code.c
@@ -2,11 +2,11 @@
 #define PI 3.14
 int main()
 {
-    int r,area;
-    scanf("%d",&r);
+    float r,area;
+    scanf("%.2f",&r);
     area=PI*r*r;
-    printf("Area: %d",area);
+    printf("Area: %.2f",area);
     return 0;
-    
+
 
 }
